[FIX] formato cotización - Min en atenciones - diagnosticos - agenda doctor
